refactor(racingcar-winner): 패키지 구조 변경 및 테스트 리팩토링
